reduce redraw signals.

+++ run.sh
@@ -21,8 +21,11 @@ while true;do
 	fdump=`timeout 2 felica_dump`
 
 	if [ ! "$fdump" ]; then
-		kill -USR2 $guicpid
-		olduser=""
+		if [ "$olduser" ]; then
+			olduser=""
+			sleep 0.3
+			kill -USR2 $guicpid
+		fi
 
 	elif [ "$fdump" = "error" ]; then
 		msg="No card Reader found"
